Reorganise file structure for dbMetaDuck
