+++ src/browser/modules/App/App.tsx
@@ -36,7 +36,7 @@ import { getErrorMessage } from 'shared/modules/commands/commandsDuck'
 import {
   findDatabaseByNameOrAlias,
   shouldAllowOutgoingConnections
-} from 'shared/modules/dbMeta/dbMetaDuck'
+} from 'shared/modules/dbMeta/selectors'
 import {
   getActiveConnection,
   getConnectionState,
