+++ src/browser/modules/DBMSInfo/DBMSInfo.tsx
@@ -38,7 +38,7 @@ import {
 } from 'browser-components/drawer/drawer-styled'
 import { DatabaseSelector } from './DatabaseSelector'
 import { getUseDb } from 'shared/modules/connections/connectionsDuck'
-import { getDatabases } from 'shared/modules/dbMeta/dbMetaDuck'
+import { getDatabases } from 'shared/modules/dbMeta/selectors'
 
 export function DBMSInfo(props: any): JSX.Element {
   const moreStep = 50
