+++ src/browser/modules/Stream/Auth/ServerStatusFrame.tsx
@@ -34,7 +34,7 @@ import {
   getActiveConnection,
   isConnected
 } from 'shared/modules/connections/connectionsDuck'
-import { shouldRetainConnectionCredentials } from 'shared/modules/dbMeta/dbMetaDuck'
+import { shouldRetainConnectionCredentials } from 'shared/modules/dbMeta/selectors'
 import { ClickToCode } from 'browser/modules/ClickToCode/index'
 
 export const ServerStatusFrame = (props: any) => {
