+++ src/browser/modules/Stream/PlayFrame.tsx
@@ -41,7 +41,7 @@ import { LAST_GUIDE_SLIDE } from 'shared/modules/udc/udcDuck'
 import { connect } from 'react-redux'
 import { GlobalState } from 'shared/globalState'
 import { inCloudEnv, inDesktop } from 'shared/modules/app/appDuck'
-import { getEdition, isEnterprise } from 'shared/modules/dbMeta/dbMetaDuck'
+import { getEdition, isEnterprise } from 'shared/modules/dbMeta/selectors'
 import { PromotionContainer, AuraPromoLink } from './styled'
 import { ThemeContext } from 'styled-components'
 import { DARK_THEME } from 'shared/modules/settings/settingsDuck'
