+++ src/browser/modules/Sync/BrowserSync.tsx
@@ -21,7 +21,7 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import TimeAgo from 'react-timeago'
-import { shouldAllowOutgoingConnections } from 'shared/modules/dbMeta/dbMetaDuck'
+import { shouldAllowOutgoingConnections } from 'shared/modules/dbMeta/selectors'
 import {
   getConnectionState,
   DISCONNECTED_STATE
