+++ src/shared/modules/commands/helpers/params.ts
@@ -21,7 +21,7 @@ import jsonic from 'jsonic'
 import { splitStringOnFirst } from 'services/commandUtils'
 import { update, replace } from 'shared/modules/params/paramsDuck'
 import { collectLambdaValues, parseLambdaStatement } from './lambdas'
-import { SYSTEM_DB } from 'shared/modules/dbMeta/dbMetaDuck'
+import { SYSTEM_DB } from 'shared/modules/dbMeta/constants'
 
 export const extractParams = (param: any) => {
   // early bail, now handled by parser
